Create first unit tests
This commit adds the first set of unit tests in the "specs" folder,
but does not add a testing "library".

Instead it makes use of the Lua Busted library which can be installed
via LuaRocks and run as a separate binary.

This can be scripted to be installed and run via Travis CI.