Remove position component from physics system
While looking for additional systems which could be tested I realised
that both the physics and collision systems were increasing the position
component by the velocity in the physics component, which meant objects
where moving twice as fast.

I decided to leave the application of the velocity to the collision system
in order to ensure objects were not clipping through other objects.

In future, collision information can then be reported back to the physics
component/system to act on in the next frame.