Integrate anim8 library into sandbox level
This commit adds Kikito's anim8 library as a submodule and implements
some basic animation in the sandbox level.

Also renames the graphics component to a sprite component as this is
closer to the truth.

Also refactors some of the code in the graphics system to better define
its purpose.

Animation code is currently very manual to set up, in future this should
be replaced with a data-driven approach, which can then be loaded and
managed by a separate class.